Load examples from a JSON file
Rather than defining examples inline, define them in a JSON file. The
example file can be loaded by using @loadExamples with a filename. See
./examples/language_examples.json for the format of the file.

The included script lzrs-gen-examples will load an examples file,
perform all of the requests contained in it against a given host:port
combination, and then re-write the file with the responses in-line.